Added Full Screen Picture View
user can now click on image in grid (of feed or me page). click the
full screen image again to go back to the tab. the me page full screen
image has two buttons. one to delete the image from the app. the other
to save the image to the public storage