Feed/Me/Posting/Database Fully Functional and Networking Improvements
I did a BUNCH of code refactoring. Firstly, got rid of almost all of
the networking classes. Using a 3rd party library to handle networking
because its cleaner and A LOT simpler. I’m having trouble posting to
the server with the new networking library.

Secondly, the database should be fully working now. There may be small
bugs because i haven’t fully testing everything but seems to work so
far with all of the updating, deleting, inserting, etc.

Thirdly, the feed and me page should be seemlessly working with the
database now. The statistics now show on the full screen view.
Hiding/disliking and liking photos via swipes works on the full screen
view of a feed image.

Lastly, posting to the server should update both the feed grid and the
me grid. Not sure if the way I coded it was the most efficient way but
it seems to work fine. Going to do more research on better
fragment/activity communication.