Refactored Image processing and database handling
Picasso has built-in image reduction which reduce the amount I had to
code manually to fix the image loading.

Also the image lists and image objects should now rely more on the
database.

Change the camera call to be in main activity and thus moved code
around a bit in the PostPicture activity.