Camera Improvements, Empty Grid Error Messages, Proper Feed Data, Image Processing
I added flash capability to the camera if your device supports it. I
cannot test on an emulator so I’m not sure if it works.

Empty grids should now show why they are empty (i.e.: no internet
connect, or no pictures in that area)

Feed grid now deletes pictures that are expired when refreshing from
the server.

Images are now scaled to a max of 1024x768 (whichever dimension
violates the constraint and keeps aspect ratio)