Better Expired Feed Data Algorithm and Improved Handling of Updating Grids/Posting
Changed the algorithm to determine if there is an expired image in the
database from O(n^2) to O(n).

Database update so all your data should get wiped.

Grids are now updated in a better way when posting to the server.

Also, when posting to the server, the feed/me grid should both show the
newly posted image right away. On the feed, it shows the actual image
file on the device now (rather than downloading from the server so we
don’t have 2 copies of the same file).