Partial Full Screen Picture Load Fix
Loading pictures in full screen view should (most of the time) load,
but still there are rare cases where the picture doesn’t load. Not sure
why.

I got rid of the Target callback that Picasso relies on however that
also means that there is no longer a progress bar when loading pictures